requirements.txt
facenet_pytorch==2.5.2
joblib==1.2.0
numpy==1.21.5
opencv_contrib_python==4.7.0.68
Pillow==9.4.0
streamlit==1.16.0
torch==1.13.1
torchvision==0.14.1
